organise components directory
