+++ resources/js/Components/Debts/AddDebt/Form/AddDebtFormAmount.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { onMounted, ref } from 'vue';
 import { store } from '@/debt.js';
-import InputError from '@/Components/InputError.vue';
+import InputError from '@/Components/Forms/InputError.vue';
 
 // props
 const props = defineProps({
