+++ resources/js/Components/GroupUsers/GroupUser.vue
@@ -2,14 +2,14 @@
 
 import { computed, onMounted, onUnmounted, ref, reactive } from 'vue';
 import { router, useForm, usePage } from '@inertiajs/vue3';
-import Modal from '@/Components/Modal.vue';
+import Modal from '@/Components/Forms/Modal.vue';
 import { Form } from '@inertiajs/vue3';
-import InputError from '@/Components/InputError.vue';
-import Controls from '@/Components/Controls.vue';
-import PrimaryButton from '@/Components/PrimaryButton.vue';
-import SecondaryButton from '@/Components/SecondaryButton.vue';
-import DangerButton from '@/Components/DangerButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import InputError from '@/Components/Forms/InputError.vue';
+import Controls from '@/Components/Misc/Controls.vue';
+import PrimaryButton from '@/Components/Misc/PrimaryButton.vue';
+import SecondaryButton from '@/Components/Misc/SecondaryButton.vue';
+import DangerButton from '@/Components/Misc/DangerButton.vue';
+import TextInput from '@/Components/Forms/TextInput.vue';
 const props = defineProps({
     group_user: {
         type: Object,
