+++ resources/js/Components/Groups/CreateGroup.vue
@@ -2,13 +2,13 @@
 <script setup>
 import { ref, onMounted } from 'vue';
 import { usePage } from '@inertiajs/vue3';
-import InputError from '@/Components/InputError.vue';
+import InputError from '@/Components/Forms/InputError.vue';
 import { Form } from '@inertiajs/vue3';
-import BigButton from '@/Components/BigButton.vue';
-import Modal from '@/Components/Modal.vue';
-import PrimaryButton from '@/Components/PrimaryButton.vue';
-import SecondaryButton from '@/Components/SecondaryButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import BigButton from '@/Components/Misc/BigButton.vue';
+import Modal from '@/Components/Forms/Modal.vue';
+import PrimaryButton from '@/Components/Misc/PrimaryButton.vue';
+import SecondaryButton from '@/Components/Misc/SecondaryButton.vue';
+import TextInput from '@/Components/Forms/TextInput.vue';
 
 const creatingGroup = ref(false);
