+++ resources/js/Components/Misc/DangerButton.vue
