+++ resources/js/Components/Misc/Slider.vue
